refactor: filterinfoView를 BeverageRecordListView로 이관
- BeverageListView에서 filterinfoView 제거
- BeverageRecordListView에 filterinfoView 추가
- filterinfoViewTapped 액션 및 로직을 BeverageRecordListViewModel로 이관
- 저당/무당 기준 모달이 BeverageRecordList 화면에서만 표시되도록 변경
