+++ HumanNeverDie/Projects/Features/BeverageRecordListFeature/Sources/BeverageRecordList/BeverageRecordListView.swift
@@ -80,7 +80,32 @@ public struct BeverageRecordListView: View {
   }
 
   private var contentView: some View {
-    BeverageListView(viewModel: viewModel.listViewModel)
+    VStack(spacing: 0) {
+      filterinfoView
+      BeverageListView(viewModel: viewModel.listViewModel)
+    }
+  }
+  
+  private var filterinfoView: some View {
+    VStack {
+      HStack(spacing: 0) {
+        Text("저당/무당 기준이 궁금하다면?")
+          .amdFont(.smallRegular)
+          .foregroundStyle(.gray60)
+        
+        AMDImage.arrowRight18.swiftUIImage
+      }
+      .frame(maxWidth: .infinity, alignment: .leading)
+      .padding(.vertical, 10)
+      .onTapGesture {
+        viewModel.handleAction(.filterinfoViewTapped)
+      }
+      .padding(.horizontal, 20)
+      
+      AMDDevider()
+    }
+    .frame(minHeight: 40, maxHeight: 40)
+    .padding(.top, 4)
   }
 
   private var sugarProgressView: some View {
