+++ HumanNeverDie/Projects/Features/BeverageRecordListFeature/Sources/BeverageRecordList/BeverageRecordListViewModel.swift
@@ -10,6 +10,7 @@ import Observation
 
 import CommonFeature
 import BeverageDomain
+import DesignSystem
 import Shared
 
 import Dependencies
@@ -34,6 +35,7 @@ public final class BeverageRecordListViewModel: ViewModelable {
     case delegateAction(BeverageListViewModel.DelegateAction?)
     case clearRoute
     case beverageLikeStatusChanged(productID: String, isLiked: Bool)
+    case filterinfoViewTapped
   }
 
   @ObservationIgnored
@@ -41,6 +43,9 @@ public final class BeverageRecordListViewModel: ViewModelable {
   
   @ObservationIgnored
   @Dependency(\.userDefaultClient) private var userDefaultClient
+  
+  @ObservationIgnored
+  @Dependency(\.alertClient) private var alertClient
   
   @ObservationIgnored
   var listViewModel: BeverageListViewModel = .init()
@@ -85,6 +90,30 @@ public final class BeverageRecordListViewModel: ViewModelable {
       
     case let .beverageLikeStatusChanged(productID, isLiked):
       syncBeverageLikeStatusFromGlobalEvent(productID: productID, isLiked: isLiked)
+    
+    case .filterinfoViewTapped:
+      let alertProperty = AMDAlertProperty(
+        title: "저당/무당 기준이 어떻게 되나요?",
+        message: """
+                 무당
+                 - 0g (제조 과정에서 당류를 첨가하지 않음)
+                 - 예: 아메리카노, 에스프레소 
+                 
+                 저당 
+                 - 액체: 100ml당 2.5g 이하
+                 - 예: 스타벅스 톨(355ml) × 2.5g = 약 8.9g 이하의 음료 모두 해당
+                 """,
+        subMessage: "아맞당은 한국 식품의약품안전처 공식 당류 표기 기준\n(식품 등의 표시·광고에 관한 법률)을 준수하고 있어요.",
+        primaryButton: .init(
+          title: "닫기",
+          type: .secondary,
+          action: {}
+        )
+      )
+      
+      Task { @MainActor in
+        await alertClient.showAlert(alertProperty)
+      }
     }
   }
