+++ HumanNeverDie/Projects/Features/BeverageRecordListFeature/Sources/Common/BeverageList/BeverageListView.swift
@@ -26,7 +26,6 @@ struct BeverageListView: View {
 
   var body: some View {
     VStack(spacing: 0) {
-      filterinfoView
       beverageFilterChipView
       beverageList
     }
@@ -37,27 +36,6 @@ struct BeverageListView: View {
     .animation(.default, value: viewModel.beverageList)
   }
   
-  private var filterinfoView: some View {
-    VStack {
-      HStack(spacing: 0) {
-        Text("저당/무당 기준이 궁금하다면?")
-          .amdFont(.smallRegular)
-          .foregroundStyle(.gray60)
-        
-        AMDImage.arrowRight18.swiftUIImage
-      }
-      .frame(maxWidth: .infinity, alignment: .leading)
-      .padding(.vertical, 10)
-      .onTapGesture {
-        viewModel.handleAction(.filterinfoViewTapped)
-      }
-      .padding(.horizontal, 20)
-      
-      AMDDevider()
-    }
-    .frame(minHeight: 40, maxHeight: 40)
-    .padding(.top, 4)
-  }
 
   private var beverageFilterChipView: some View {
     HStack(spacing: 4) {
